New sample: text
I'm kind of tilted. I wasn't able to make this work. Renderdoc shows
that bitmaps are properly loaded into atlas but on second frame font
atlas gets messed up. I don't know why this is happening. I thought
maybe I was overwriting the atlas from somewhere but it doesn't seem
to be the case.

Anyway, I'll take a break (not sure how long it will be) and we'll
continue with coding.